refactor(auto_labeling_3d): add BaseEnsembleModel
