+++ tools/auto_labeling_3d/filter_objects/config/ensemble_base.py
@@ -34,7 +34,7 @@ bevfusion_pipeline = [
 filter_pipelines = dict(
     type="Ensemble",
     config=dict(
-        type="EnsembleModel",
+        type="NMSEnsembleModel",
         ensemble_setting=dict(
             weights=[1.0, 1.0],
             iou_threshold=0.55,
